refactor: extract common react code into client/react-base package
