+++ packages/client/react-base/src/hooks/index.ts
@@ -1,1 +1,2 @@
+export * from "./useCrossmint";
 export * from "./useWallet";
