+++ packages/client/react-base/src/index.ts
@@ -1,2 +1,3 @@
 export * from "./hooks";
 export * from "./providers";
+export * from "./types";
