+++ packages/client/ui/react-native/package.json
@@ -16,8 +16,7 @@
     "files": ["dist", "LICENSE"],
     "scripts": {
         "build": "cross-env NODE_OPTIONS='--max-old-space-size=8192' tsup",
-        "dev": "cross-env NODE_OPTIONS='--max-old-space-size=8192' tsup --watch",
-        "test:vitest": "vitest run"
+        "dev": "cross-env NODE_OPTIONS='--max-old-space-size=8192' tsup --watch"
     },
     "dependencies": {
         "@crossmint/common-sdk-base": "workspace:*",
