+++ packages/client/ui/react-native/src/hooks/index.ts
@@ -1,4 +1,2 @@
-import { useWallet } from "@crossmint/client-sdk-react-base";
-
+export { useCrossmint, useWallet } from "@crossmint/client-sdk-react-base";
 export * from "./useCrossmint";
-export { useWallet };
